# Add collector local cache of abs_file