As per [Issue 695](https://bitbucket.org/ned/coveragepy/issues/625/lstat-dominates-in-the-case-of-small), for Hypothesis's weird idiosyncratic usage pattern of coverage, the cost of calling `abs_path` in `save_data` turns out to be quite significant. This implements the simplest possible fix for that, which is that it adds caching to the function.

I wasn't terribly comfortable with adding module-level state, so I implemented a less invasive way of doing this, which is that I added a cache that was local to the collector.

I don't have any great way of benchmarking this, but informally to see the effect of the change I ran one of Hypothesis's test files (`test_testdecorators.py`, which is basically just a smoke test of some typical usage examples). Prior to this change it took about 14 seconds, and after this change it took 11 seconds. Given that most of the time it's spending is probably not in coverage, that's a fairly dramatic improvement!

ETA: On rerunning the same experiment (with the system under slightly different load, so slightly different numbers) I get the following:

* Without using coverage  ~7.5s
* With using coverage and without this change ~13s
* With using coverage and with this change: ~10s.

So for Hypothesis usage at least this change is probably a huge win performance-wise (but I understand Hypothesis usage is weird. Also this is on a system with unusually slow disk access).