@nedbat 2018-05-14
Applied as https://bitbucket.org/ned/coveragepy/commits/44f0e230c68e611c9edfdf28fbdad73dd502afe5

Thanks!